move check_shape_adj_op
