+++ src/mrinufft/operators/interfaces/tfnufft.py
@@ -2,7 +2,6 @@
 
 import numpy as np
 
-from mrinufft.operators.interfaces.utils import check_shape_adj_op
 from ..base import FourierOperatorBase, with_tensorflow
 from mrinufft._utils import proper_trajectory
 
@@ -110,7 +109,7 @@ class MRITensorflowNUFFT(FourierOperatorBase):
         -------
         Tensor
         """
-        check_shape_adj_op(self, data)
+        self.check_shape_adj_op(self, data)
         if self.uses_density:
             data_d = data * self.density
         else:
