+++ src/mrinufft/operators/interfaces/utils/__init__.py
@@ -4,7 +4,6 @@ from .utils import (
     check_error,
     check_size,
     sizeof_fmt,
-    check_shape_adj_op,
 )
 
 from .gpu_utils import (
@@ -20,7 +19,6 @@ from .gpu_utils import (
 __all__ = [
     "check_error",
     "check_size",
-    "check_shape_adj_op",
     "sizeof_fmt",
     "get_maxThreadBlock",
     "CUPY_AVAILABLE",
