+++ src/mrinufft/operators/interfaces/utils/utils.py
@@ -35,26 +35,3 @@ def check_size(array_like, shape):
     if np.prod(array_like.shape) != np.prod(shape):
         raise ValueError(f"Expected array with {shape}, got {array_like.shape}.")
 
-
-def check_shape_adj_op(self_, image):
-    """Validate that the shape of the provided image matches the expected shape.
-
-    This validation is defined by the operator during initialization.
-
-    Parameters
-    ----------
-    image : np.ndarray or Tensor
-
-    Returns
-    -------
-    None
-        This function does not return any value. It raises a ValueError if the
-        image shape does not match the expected shape.
-    """
-    image_samples = image.shape[-1]
-
-    if image_samples != self_.n_samples:
-        raise ValueError(
-            f"Image shape {image_samples} is not compatible "
-            f"with the operator shape {self_.n_samples}"
-        )
